feat: fix handleChange
- introduced a white screen of death on submit
- validates based on rule
