--- src/components/LayoutSelector/ButtonsFooter/styles.ts
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-import { Space } from "antd";
-
-const FlexBox = styled(Space)`
-  padding: 24px;
-  margin: 0 auto;
-  min-height: 95%;
-  width: 80%;
-`;
-
-export default FlexBox;
