+++ src/components/LayoutSelector/Information/index.tsx
@@ -1,6 +1,5 @@
 import { FC } from "react";
 import { Typography, Row, Col } from "antd";
-import LayoutSelectorGlobalStyled from "../styles";
 
 const { Paragraph } = Typography;
 
@@ -11,27 +10,24 @@ const Information: FC = () => {
     "Any text block can contain a background image or color",
   ];
   return (
-    <>
-      <LayoutSelectorGlobalStyled />
-      <Row
-        gutter={[12, 48]}
-        justify="space-between"
-        style={{
-          fontSize: "initial",
-        }}
-      >
-        <Col style={{ width: "100%" }}>
-          <Paragraph>
-            When choosing your layout there are several things you can customize
-          </Paragraph>
-          <ul>
-            {bullets.map((bullet) => (
-              <li>{bullet}</li>
-            ))}
-          </ul>
-        </Col>
-      </Row>
-    </>
+    <Row
+      gutter={[12, 48]}
+      justify="space-between"
+      style={{
+        fontSize: "initial",
+      }}
+    >
+      <Col style={{ width: "100%" }}>
+        <Paragraph>
+          When choosing your layout there are several things you can customize
+        </Paragraph>
+        <ul>
+          {bullets.map((bullet) => (
+            <li key={bullet}>{bullet}</li>
+          ))}
+        </ul>
+      </Col>
+    </Row>
   );
 };
